+++ js/data.js
@@ -1,0 +1,71 @@
+const getDataArrayMiniatures = () => {
+  const DESCRIPTIONS = [
+    'Вид на озеро при отеле.',
+    'Указатель ведущий к пляжу.',
+    'Вид на пляж и лазурный океан.',
+    'Маша на берегу океана.',
+    'Весёлый обед.',
+    'Спортивный автомобиль моей мечты.',
+    'Клубничный завтрак.',
+    'Прохладный морс из свежей смородины.',
+    'Удачное фото на берегу океана с пролетающим кукурузником.',
+    'Коллекция летней обуви.',
+    'Тропа к океану.',
+    'Белая ауди.',
+    'На ужин: мясо с овощами.',
+    'Суши-котик.',
+    'Тапки железного человека.',
+    'Вид с борта самолёта.',
+    'Выступление хора.',
+    'Ретро автомобиль.',
+    'Очень удобные тапочки с подстветкой.',
+    'Отель в восточных масивах среди высоких пальм.',
+    'Пэпэшный обед.',
+    'Незабываем закат на берегу океана.',
+    'Краб сухопутный тритаился.',
+    'Концерт любомого исполнителя.',
+    'Бегемот хотел нас съесть'
+  ];
+
+  const COMMENTS = [
+    'Всё отлично!',
+    'В целом всё неплохо. Но не всё.',
+    'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+    'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+    'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+    'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+  ];
+
+  const NAMES = [
+    'Дарья',
+    'Артём',
+    'Ольга',
+    'Алексей',
+    'Александр',
+    'Гульнара',
+    'Дмитрий',
+    'Еремей',
+    'Тимофей',
+    'Кутуп',
+    'Полина',
+    'Валерия',
+    'Клавдия',
+    'Патрик',
+    'Адам',
+    'Ева',
+    'Маргарита',
+    'Есения',
+    'Мира',
+    'Екатерина',
+    'Мария',
+    'Юлия',
+    'Оксана',
+    'Михаил',
+    'Клементий'
+  ];
+  return { DESCRIPTIONS, COMMENTS, NAMES };
+};
+
+export { getDataArrayMiniatures };
+
+
